refactor: 파일 분리
테스트 및 가독성을 위한 validate 메서드들을 별도로 분리
