+++ src/main/java/racingcar/Application.java
@@ -7,9 +7,16 @@ import java.util.Arrays;
 
 public class Application {
     public static void main(String[] args) {
+
+        Validation validation = new Validation();
+
         System.out.println("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)");
         String input = Console.readLine();
-        validateInput(input);
+        validation.validateInput(input);
+
+        String[] carNames = input.split(",");
+        validation.validateCarNames(carNames);
+        validation.validateLength(carNames);
 
         System.out.println("시도할 회수는 몇회인가요?");
         int tryCount = 0;
@@ -18,9 +25,6 @@ public class Application {
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException();
         }
-
-        String[] carNames = input.split(",");
-        validateLength(carNames);
 
         int[] raceDistance = new int[carNames.length];
         int winnerDistance = 0;
@@ -51,6 +55,7 @@ public class Application {
 
     }
 
+
     public static boolean goOrStop() {
         if (Randoms.pickNumberInRange(0, 9) >= 4) {
             return true;
@@ -59,23 +64,6 @@ public class Application {
         }
     }
 
-    public static void validateInput(String input) {
-        //사용자의 입력이 "," 혹은 알파벳이 아니면 IllegalArgumentException 발생
-        for (int i = 0; i < input.length(); i++) {
-            if (!Character.isAlphabetic(input.charAt(i)) && input.charAt(i) != ',') {
-                throw new IllegalArgumentException();
-            }
-        }
-    }
-
-    public static void validateLength(String[] carNames) {
-        //자동차 이름이 5자를 넘어가면 IllegalArgumentException 발생
-        for (int i = 0; i < carNames.length; i++) {
-            if (carNames[i].length() > 5) {
-                throw new IllegalArgumentException();
-            }
-        }
-    }
 
     public static void raceProceed(String[] carNames, int[] raceDistance, String[] raceDistanceByHyphen) {
         for (int j = 0; j < carNames.length; j++) {
